ethernet_pcap: Move NE2000 pcap code to PcapEthernetConnection
This change separates the pcap code out of the NE2000 code to its own
EthernetConnection backend, and modifies the NE2000 to use the
EthernetConnection abstraction.
