+++ src/hardware/ne2000.cpp
@@ -1,9 +1,5 @@
 #include "config.h"
 
-#if defined(WIN32)
-  #define HAVE_REMOTE
-#endif
-
 #include "dosbox.h"
 #include <string.h>
 #include <stdio.h>
@@ -51,33 +47,10 @@
 
 #include "ne2000.h"
 
+#include "ethernet.h"
+EthernetConnection* ethernet = nullptr;
+
 static void NE2000_TX_Event(Bitu val);
-
-#ifdef C_PCAP
-
-#include "pcap.h"
-// Handle to WinPCap device
-pcap_t *adhandle = 0;
-
-#ifdef WIN32
-// DLL loading
-#define pcap_sendpacket(A,B,C)			PacketSendPacket(A,B,C)
-#define pcap_close(A)					PacketClose(A)
-#define pcap_freealldevs(A)				PacketFreealldevs(A)
-#define pcap_open(A,B,C,D,E,F)			PacketOpen(A,B,C,D,E,F)
-#define pcap_next_ex(A,B,C)				PacketNextEx(A,B,C)
-#define pcap_findalldevs_ex(A,B,C,D)	PacketFindALlDevsEx(A,B,C,D)
-
-int (*PacketSendPacket)(pcap_t *, const u_char *, int) = 0;
-void (*PacketClose)(pcap_t *) = 0;
-void (*PacketFreealldevs)(pcap_if_t *) = 0;
-pcap_t* (*PacketOpen)(char const *,int,int,int,struct pcap_rmtauth *,char *) = 0;
-int (*PacketNextEx)(pcap_t *, struct pcap_pkthdr **, const u_char **) = 0;
-int (*PacketFindALlDevsEx)(char *, struct pcap_rmtauth *, pcap_if_t **, char *) = 0;
-
-#endif
-
-#endif
 
 //Never completely fill the ne2k ring so that we never
 // hit the unclear completely full buffer condition.
@@ -304,9 +277,7 @@ bx_ne2k_c::write_cr(uint32_t value)
     // Send the packet to the system driver
 	/* TODO: Transmit packet */
     //BX_NE2K_THIS ethdev->sendpkt(& BX_NE2K_THIS s.mem[BX_NE2K_THIS s.tx_page_start*256 - BX_NE2K_MEMSTART], BX_NE2K_THIS s.tx_bytes);
-#ifdef C_PCAP
-	pcap_sendpacket(adhandle,&s.mem[s.tx_page_start*256 - BX_NE2K_MEMSTART], s.tx_bytes);
-#endif
+    	ethernet->SendPacket(&s.mem[s.tx_page_start*256 - BX_NE2K_MEMSTART], s.tx_bytes);
 	// some more debug
 	if (BX_NE2K_THIS s.tx_timer_active) {
       BX_PANIC(("CR write, tx timer still active"));
@@ -1473,25 +1444,15 @@ static void NE2000_TX_Event(Bitu val) {
 }
 
 static void NE2000_Poller(void) {
-#ifdef C_PCAP
-	int res;
-	struct pcap_pkthdr *header;
-	u_char *pkt_data;
-//#if 0
-	while((res = pcap_next_ex( adhandle, &header, (const u_char **)&pkt_data)) > 0) {
+	ethernet->GetPackets([](const uint8_t* packet, int len) {
 		//LOG_MSG("NE2000: Received %d bytes", header->len);
-		
+
 		// don't receive in loopback modes
 		if((theNE2kDevice->s.DCR.loop == 0) || (theNE2kDevice->s.TCR.loop_cntl != 0))
 			return;
-		theNE2kDevice->rx_frame(pkt_data, header->len);
-	}
-//#endif
-#endif
+		theNE2kDevice->rx_frame(packet, len);
+	});
 }
-#ifdef WIN32
-#include <windows.h>
-#endif
 
 extern std::string niclist;
 
@@ -1516,67 +1477,12 @@ public:
 			return;
 		}
 
-#ifndef C_PCAP
-		LOG_MSG("pcap support not enabled, disabling ne2000");
-		load_success = false;
-		return;
-#endif
-
-#ifdef C_PCAP
-#ifdef WIN32
-/*		
-		int (*PacketSendPacket)(pcap_t *, const u_char *, int);
-		void (*PacketClose)(pcap_t *);
-		void (*PacketFreealldevs)(pcap_if_t *);
-		pcap_t* (*PacketOpen)(char const *,int,int,int,struct pcap_rmtauth *,char *);
-		int (*PacketNextEx)(pcap_t *, struct pcap_pkthdr **, const u_char **);
-		int (*PacketFindALlDevsEx)(char *, struct pcap_rmtauth *, pcap_if_t **, char *);
-*/
-		// init the library
-		HINSTANCE pcapinst;
-		pcapinst = LoadLibrary("WPCAP.DLL");
-		if(pcapinst==NULL) {
-            niclist = "WinPcap has to be installed for the NE2000 to work.";
-			LOG_MSG(niclist.c_str());
+		ethernet = OpenEthernetConnection("pcap");
+		if(!ethernet)
+		{
 			load_success = false;
 			return;
 		}
-		FARPROC psp;
-		
-		psp = GetProcAddress(pcapinst,"pcap_sendpacket");
-		if(!PacketSendPacket) PacketSendPacket =
-			(int (__cdecl *)(pcap_t *,const u_char *,int))psp;
-		
-		psp = GetProcAddress(pcapinst,"pcap_close");
-		if(!PacketClose) PacketClose =
-			(void (__cdecl *)(pcap_t *)) psp;
-		
-		psp = GetProcAddress(pcapinst,"pcap_freealldevs");
-		if(!PacketFreealldevs) PacketFreealldevs =
-			(void (__cdecl *)(pcap_if_t *)) psp;
-
-		psp = GetProcAddress(pcapinst,"pcap_open");
-		if(!PacketOpen) PacketOpen =
-			(pcap_t* (__cdecl *)(char const *,int,int,int,struct pcap_rmtauth *,char *)) psp;
-
-		psp = GetProcAddress(pcapinst,"pcap_next_ex");
-		if(!PacketNextEx) PacketNextEx = 
-			(int (__cdecl *)(pcap_t *, struct pcap_pkthdr **, const u_char **)) psp;
-
-		psp = GetProcAddress(pcapinst,"pcap_findalldevs_ex");
-		if(!PacketFindALlDevsEx) PacketFindALlDevsEx =
-			(int (__cdecl *)(char *, struct pcap_rmtauth *, pcap_if_t **, char *)) psp;
-
-		if(PacketFindALlDevsEx==0 || PacketNextEx==0 || PacketOpen==0 || 
-			PacketFreealldevs==0 || PacketClose==0 || PacketSendPacket==0) {
-            niclist = "Incorrect or non-functional WinPcap version.";
-			LOG_MSG(niclist.c_str());
-			load_success = false;
-			return;
-		}
-
-#endif
-#endif
 
 		// get irq and base
 		Bitu irq = (Bitu)section->Get_int("nicirq");
@@ -1605,109 +1511,6 @@ public:
 			mac[4]=macint[4]; mac[5]=macint[5];
 		}
 
-#ifdef C_PCAP
-		// find out which pcap device to use
-		const char* realnicstring=section->Get_string("realnic");
-		pcap_if_t *alldevs;
-		pcap_if_t *currentdev = NULL;
-		char errbuf[PCAP_ERRBUF_SIZE];
-		unsigned int userdev;
-#ifdef WIN32
-		if (pcap_findalldevs_ex(PCAP_SRC_IF_STRING, NULL, &alldevs, errbuf) == -1)
-#else
-		if (pcap_findalldevs(&alldevs, errbuf) == -1)
-#endif
- 		{
-            niclist = "Cannot enumerate network interfaces: "+std::string(errbuf);
-			LOG_MSG("%s", niclist.c_str());
-			load_success = false;
-			return;
-		}
-        {
-            Bitu i = 0;
-            niclist = "Network Interface List\n-------------------------------------------------------------\n";
-            for(currentdev=alldevs; currentdev!=NULL; currentdev=currentdev->next) {
-                const char* desc = "no description";
-                if(currentdev->description) desc=currentdev->description;
-                i++;
-                niclist+=(i<10?"0":"")+std::to_string(i)+" "+currentdev->name+"\n    ("+desc+")\n";
-            }
-        }
-		if (!strcasecmp(realnicstring,"list")) {
-			// print list and quit
-            std::istringstream in(("\n"+niclist+"\n").c_str());
-            if (in)	for (std::string line; std::getline(in, line); )
-                LOG_MSG("%s", line.c_str());
-			pcap_freealldevs(alldevs);
-			load_success = false;
-			return;
-		} else if(1==sscanf(realnicstring,"%u",&userdev)) {
-			// user passed us a number
-			Bitu i = 0;
-			currentdev=alldevs;
-			while(currentdev!=NULL) {
-				i++;
-				if(i==userdev) break;
-				else currentdev=currentdev->next;
-			}
-		} else {
-			// user might have passed a piece of name
-			for(currentdev=alldevs; currentdev!=NULL; currentdev=currentdev->next) {
-				if(strstr(currentdev->name,realnicstring)) {
-					break;
-				}else if(currentdev->description!=NULL &&
-					strstr(currentdev->description,realnicstring)) {
-					break;
-				}
-			}
-		}
-
-		if(currentdev==NULL) {
-			LOG_MSG("Unable to find network interface - check realnic parameter\n");
-			load_success = false;
-			pcap_freealldevs(alldevs);
-			return;
-		}
-		// print out which interface we are going to use
-        const char* desc = "no description"; 
-		if(currentdev->description) desc=currentdev->description;
-		LOG_MSG("Using Network interface:\n%s\n(%s)\n",currentdev->name,desc);
-		
-		// attempt to open it
-#ifdef WIN32
-		if ( (adhandle= pcap_open(
-			currentdev->name, // name of the device
-            65536,            // portion of the packet to capture
-                              // 65536 = whole packet 
-            PCAP_OPENFLAG_PROMISCUOUS,    // promiscuous mode
-            3000,             // read timeout
-            NULL,             // authentication on the remote machine
-            errbuf            // error buffer
-            ) ) == NULL)
-#else
-		/*pcap_t *pcap_open_live(const char *device, int snaplen,
-               int promisc, int to_ms, char *errbuf)*/
-		if ( (adhandle= pcap_open_live(
-			currentdev->name, // name of the device
-            65536,            // portion of the packet to capture
-                              // 65536 = whole packet 
-            true,    // promiscuous mode
-            3000,             // read timeout
-            errbuf            // error buffer
-            ) ) == NULL)
-
-#endif        
-        {
-				LOG_MSG("\nUnable to open the interface: %s.", errbuf);
-        	pcap_freealldevs(alldevs);
-			load_success = false;
-			return;
-		}
-		pcap_freealldevs(alldevs);
-#ifndef WIN32
-		pcap_setnonblock(adhandle,1,errbuf);
-#endif
-#endif
 		// create the bochs NIC class
 		theNE2kDevice = new bx_ne2k_c ();
 		memcpy(theNE2kDevice->s.physaddr, mac, 6);
@@ -1728,10 +1531,8 @@ public:
 	}	
 	
 	~NE2K() {
-#ifdef C_PCAP
-		if(adhandle) pcap_close(adhandle);
-		adhandle=0;
-#endif
+		if(ethernet != 0) delete ethernet;
+		ethernet=0;
 		if(theNE2kDevice != 0) delete theNE2kDevice;
 		theNE2kDevice=0;
 		TIMER_DelTickHandler(NE2000_Poller);
