+++ src/misc/Makefile.am
@@ -3,4 +3,5 @@ AM_CPPFLAGS = -I$(top_srcdir)/include "-DRESDIR=\"$(resdir)\""
 resdir = $(datarootdir)/dosbox-x
 
 noinst_LIBRARIES = libmisc.a
-libmisc_a_SOURCES = cross.cpp ethernet.cpp messages.cpp programs.cpp setup.cpp support.cpp regionalloctracking.cpp shiftjis.cpp iconvpp.cpp
+noinst_HEADERS = ethernet_pcap.h
+libmisc_a_SOURCES = cross.cpp ethernet.cpp ethernet_pcap.cpp messages.cpp programs.cpp setup.cpp support.cpp regionalloctracking.cpp shiftjis.cpp iconvpp.cpp
